Merge develop for v4.0.1

+++ php/commands/src/CLI_Command.php
@@ -286,6 +286,7 @@ class CLI_Command extends EE_Command {
 			$download_url = $newest['package_url'];
 			$md5_url      = str_replace( '.phar', '.phar.md5', $download_url );
 		}
+		EE::get_runner()->check_requirements();
 		EE::log( sprintf( 'Downloading from %s...', $download_url ) );
 		$temp    = \EE\Utils\get_temp_dir() . uniqid( 'ee_', true ) . '.phar';
 		$headers = array();
@@ -493,21 +494,27 @@ class CLI_Command extends EE_Command {
 	 */
 	public function self_uninstall( $args, $assoc_args ) {
 
+		if ( ! EE::get_runner()->check_requirements( false ) ) {
+			EE::error( 'Unable to proceed with uninstallation. Seems there is a dependency down.', false );
+			die;
+		}
+
 		EE::confirm( "Are you sure you want to remove EasyEngine and all its sites(along with their data)?\nThis is an irreversible action. No backup will be kept.", $assoc_args );
 
 		EE::exec( 'docker rm -f $(docker ps -aqf label=org.label-schema.vendor="EasyEngine")' );
-		$home = Utils\get_home_dir();
-		EE::exec( "rm -rf $home/.ee/" );
+		EE::exec( 'docker network prune -f $(docker network ls -f "label=org.label-schema.vendor=EasyEngine")' );
+		EE::exec( 'docker volume rm -f $(docker volume ls -f "label=org.label-schema.vendor=EasyEngine" -q)' );
+		EE::exec( 'docker image rm $(docker image ls -f "label=org.label-schema.vendor=EasyEngine" -q)' );
 
 		$records = Site::all( [ 'site_fs_path' ] );
 
+		$fs = new Filesystem();
 		if ( ! empty( $records ) ) {
 			$sites_paths = array_column( $records, 'site_fs_path' );
-			$fs = new Filesystem();
 			$fs->remove( $sites_paths );
 		}
 
-		EE::exec( "rm -rf " . EE_ROOT_DIR );
+		$fs->remove( EE_ROOT_DIR );
 
 		if ( Utils\inside_phar() ) {
 			unlink( realpath( $_SERVER['argv'][0] ) );
